+++ plugins/ikfastsolvers/ikfastsolver.cpp
@@ -733,7 +733,7 @@ private:
                 for(int i = 0; i < iksol.GetDOF(); ++i) {
                     vravesol.at(i) = (dReal)sol[i];
                 }
-                vdists.push_back(make_pair(vdists.size(),_configdist2(probot,vravesol,q0)));
+                vdists.push_back(make_pair(vdists.size(),_ComputeGeometricConfigDistSqr(probot,vravesol,q0,true)));
             }
 
             std::stable_sort(vdists.begin(),vdists.end(),SortSolutionDistances);
@@ -815,7 +815,7 @@ private:
                 // if all the solutions are worse than the best, then ignore everything
                 vravesols2.reserve(vravesols.size());
                 FOREACH(itravesol, vravesols) {
-                    d = _configdist2(probot,itravesol->first,boost::get<1>(freeq0check));
+                    d = _ComputeGeometricConfigDistSqr(probot,itravesol->first,boost::get<1>(freeq0check));
                     if( !(bestsolution.dist <= d) ) {
                         vravesols2.push_back(*itravesol);
                     }
@@ -828,7 +828,7 @@ private:
         }
         else {
             if( boost::get<1>(freeq0check).size() == vravesol.size() ) {
-                d = _configdist2(probot,vravesol,boost::get<1>(freeq0check));
+                d = _ComputeGeometricConfigDistSqr(probot,vravesol,boost::get<1>(freeq0check));
                 if( bestsolution.dist <= d ) {
                     return IKRA_Reject;
                 }
@@ -946,7 +946,7 @@ private:
         size_t index = 0;
         FOREACH(itikreturn, listlocalikreturns) {
             if( (int)boost::get<1>(freeq0check).size() == _nTotalDOF ) {
-                d = _configdist2(probot,(*itikreturn)->_vsolution,boost::get<1>(freeq0check));
+                d = _ComputeGeometricConfigDistSqr(probot,(*itikreturn)->_vsolution,boost::get<1>(freeq0check));
                 if( !(bestsolution.dist <= d) ) {
                     bestsolution.ikreturn = *itikreturn;
                     bestsolution.dist = d;
@@ -1127,17 +1127,26 @@ private:
         return true;
     }
 
-    dReal _configdist2(RobotBasePtr probot, const vector<dReal>& q1, const vector<dReal>& q2) const
+    /// \brief configuraiton distance
+    ///
+    /// \param bNormalizeRevolute if true, then compute difference mod 2*PI
+    dReal _ComputeGeometricConfigDistSqr(RobotBasePtr probot, const vector<dReal>& q1, const vector<dReal>& q2, bool bNormalizeRevolute=false) const
     {
         vector<dReal> q = q1;
         probot->SubtractActiveDOFValues(q,q2);
         vector<dReal>::iterator itq = q.begin();
+        std::vector<uint8_t>::const_iterator itrevolute = _vjointrevolute.begin();
         dReal dist = 0;
         FOREACHC(it, probot->GetActiveDOFIndices()) {
             KinBody::JointPtr pjoint = probot->GetJointFromDOFIndex(*it);
             dReal fweight = pjoint->GetWeight(*it-pjoint->GetDOFIndex());
+            // have to do the
+            if( bNormalizeRevolute && *itrevolute ) {
+                *itq = utils::NormalizeCircularAngle(*itq, -PI, PI);
+            }
             dist += *itq**itq * fweight * fweight;
-            itq++;
+            ++itq;
+            ++itrevolute;
         }
         return dist;
     }